move vuln parsing logic to common
