+++ packages/repocop/src/utils.test.ts
@@ -1,10 +1,6 @@
-import type { NonEmptyArray, RepocopVulnerability, Repository } from './types';
-import {
-	isProduction,
-	stringToSeverity,
-	toNonEmptyArray,
-	vulnSortPredicate,
-} from './utils';
+import type { RepocopVulnerability } from 'common/src/types';
+import type { NonEmptyArray, Repository } from './types';
+import { isProduction, toNonEmptyArray, vulnSortPredicate } from './utils';
 
 describe('isProduction', () => {
 	test('should return correct values for prod and non-prod repos', () => {
@@ -39,18 +35,6 @@ describe('Failure on empty arrays', () => {
 		expect(() => toNonEmptyArray(nonEmptyArray)).not.toThrow();
 		expect(toNonEmptyArray(nonEmptyArray)).toEqual(nonEmptyArray);
 		expect(toNonEmptyArray(nonEmptyArray)).toEqual(typedNonEmptyArray);
-	});
-});
-
-describe('stringToSeverity', () => {
-	test('should return unknown if it is passed an unexpected string', () => {
-		expect(stringToSeverity('foo')).toBe('unknown');
-	});
-	test('should return the correct severity for valid inputs', () => {
-		expect(stringToSeverity('low')).toBe('low');
-		expect(stringToSeverity('medium')).toBe('medium');
-		expect(stringToSeverity('high')).toBe('high');
-		expect(stringToSeverity('critical')).toBe('critical');
 	});
 });
